+++ cloudify_cli/VERSION
@@ -1,6 +1,6 @@
 {
-  "date": "", 
+  "date": "",
   "commit": "",
-  "version": "3.2.0", 
+  "version": "3.2.0",
   "build": ""
 }
